feat(bc)!: follow behavior-contracts 0.11.9 — the wire carries numbers natively and is never cloned
0.11.9 lands bc#229: `WireValue` splits `Num(text)` into native `Int(i64)` / `Float(f64)`, keys and
strings become `Cow<'static, str>` (a generated column name rides borrowed, so it is not re-allocated
per row), and the rust probes become free functions that CONSUME the wire instead of cloning it. The
generated modules shrank accordingly - rust sqlite went from 82 `clone()` / 54 `to_string()` to 24 / 23.

Every pin is 0.11.9 (npm, rust workspace + orm_bench, go, python; the php port re-vendored at 006dc3d),
and all five runtimes follow the new API. Four places in OUR code had to change, all of them the same
mistake in different languages - a read that handed bc a float where the column is an `Int`:

- `sqldb.go` widened every scanned `int64` to `float64` to mirror a JS number. With int and float now
  distinct wire kinds, that made EVERY typed read fail `expected int, got N`. An INTEGER column stays
  an int64.
- `lib.rs` re-exported the wire types but not the probe functions, so a generated module
  (`use litedbmodel_runtime::*`, from `--shared-types-import`) could not resolve `probe_list_at` and
  friends. What a generated module calls is part of the import contract.
- `livedb.rs`'s PostgreSQL and MySQL row decoders built `WireValue::Str(String)` and
  `Vec<(String, _)>`; both take a `Cow` now. Only the sqlite target was being built, which is how this
  stayed hidden until the live legs ran.
- `leaf_transport.go` / `grouping.rs` / `leaves.rs` read numbers as text. `parseWireNum` is deleted
  outright - there is nothing left to parse - and the key identity renders straight off the native
  number instead of parsing and re-formatting it per row.

Tests follow the same fact: an INTEGER column reads back as bc's `int`. rust 61 passed, go green,
TypeScript 1227 passed.

+++ go/litedbmodel_runtime/sqldb.go
@@ -224,7 +224,10 @@ func scanValue(v any) bc.Value {
 	case nil:
 		return nil
 	case int64:
-		return float64(t) // JS number (better-sqlite3 integer column → plain JSON number)
+		// An INTEGER column stays an int64. It used to be widened to float64 to mirror a JS number, but
+		// the wire carries int and float as DISTINCT kinds, so a column declared `Int` cannot be satisfied
+		// by a float — widening here made every typed read fail `expected int, got N`.
+		return t
 	case float64:
 		return t
 	case bool:
